requirements.txt
hyperliquid-python-sdk==0.20.1
numpy>=1.21.0,<2.0.0
pandas>=1.5.0,<3.0.0
scikit-learn>=1.2.0,<2.0.0
TA-Lib>=0.4.25
requests>=2.28.0
python-dotenv>=0.19.0
schedule>=1.2.0
matplotlib>=3.5.0
seaborn>=0.11.0
joblib>=1.2.0
yfinance>=0.2.0
websocket-client>=1.4.0
aiohttp>=3.8.0
textblob>=0.17.0
vaderSentiment>=3.3.0
loguru>=0.6.0
asyncio>=3.4.3
python-telegram-bot>=20.0
psutil>=5.9.0